4.21: blogilistan laajennus

+++ osa4/blogilista/controllers/blogs.js
@@ -43,7 +43,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const userid = await User.findById(decodedToken.id)
+
+  if ( blog.user.toString() === userid.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+  }
+
   response.status(204).end()
 })
